Minor bug fixes; automatic input data conversion; updated docs

+++ include/nodes/datatypes/decimaldata.hpp
@@ -23,7 +23,7 @@ namespace nitro {
         [[nodiscard]] QString getDescription() const override;
 
     private:
-        inline static const QString name_ = "Double";
+        inline static const QString name_ = "Decimal";
         inline static const QString id_ = "Decimal";
         inline static const QColor baseColor_ = {161, 161, 161};
     };
